+++ smooth_progress/models.py
@@ -1,5 +1,4 @@
 from math import floor
-from typing import NoReturn
 from .exceptions import ProgressBarClosedError
 
 class ProgressBar:
@@ -7,13 +6,13 @@ class ProgressBar:
     this class. Once the project is more completed, attributes will be hidden and
     available through getter/setter method pairs.
     """
-    def __enter__(self: object):
+    def __enter__(self: object) -> None:
         return self
 
-    def __exit__(self: object, t, val, tb):
+    def __exit__(self: object, t, val, tb) -> None:
         del self
 
-    def __init__(self: object, limit: int = 100, show_percent: bool = True):
+    def __init__(self: object, limit: int = 100, show_percent: bool = True) -> None:
         """
         :param limit: int, optional, default 100.
         :param show_percent: bool, optional, default True.
@@ -39,7 +38,7 @@ class ProgressBar:
         else:
             return False
 
-    def increment(self: object) -> NoReturn:
+    def increment(self: object) -> None:
         """Increments the progress and updates the display to reflect the new value. If
         this incrementation takes the progress to the pre-defined limit, closes the
         ProgressBar from mutability.
@@ -78,7 +77,16 @@ class ProgressBar:
             self.opened = True
             return True
 
-    def __update(self: object, completion: int, percentage: int) -> NoReturn:
+    def show(self: object, end: str = "\n") -> None:
+        """Display the current state of the bar, with the end character determined by
+        the call.
+
+        :param end: The end character, by default a new line. Only control characters
+                    are recommended, but any string can be passed.
+        """
+        print(self.state, end=end, flush=True)
+
+    def __update(self: object, completion: int, percentage: int) -> None:
         """Hidden method to update the state of the bar with new values. Should only be
         called from within the class itself.