+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Unreleased
 
+- Added `ProgressBar.show()` method, which displays the current state of the bar. (@MurdoMaclachlan)
 - `ProgressBar.close()`, `ProgressBar.interrupt()` and `ProgressBar.open()` now return a boolean value indicating if they had any effect. (@MurdoMaclachlan)
+- Functions with no return value now use `None` type hinting instead of `NoReturn`, as per the standard. (@MurdoMaclachlan)
 - Fixed a typo in the README (@MurdoMaclachlan)
 
 ### 0.1.0